+++ packages/server-python/src/rfnry_chat_server/analytics/__init__.py
@@ -1,3 +1,0 @@
-from rfnry_chat_server.analytics.collector import AnalyticsEvent, AssistantAnalytics, OnAnalyticsCallback
-
-__all__ = ["AnalyticsEvent", "AssistantAnalytics", "OnAnalyticsCallback"]